+++ src/main/java/nom/tam/fits/HeaderCard.java
@@ -512,12 +512,17 @@ public class HeaderCard implements CursorValue<String>, Cloneable {
      * @see                        #set(String, String, String, Class)
      */
     private HeaderCard(String key, String value, String comment, Class<?> type) throws HeaderCardException {
-        if (value == null) {
-            if (key == null || key.equals(Standard.COMMENT.key()) || key.equals(Standard.BLANKS.key())
-                    || key.equals(Standard.HISTORY.key())) {
-                // Force comment
-                type = null;
+        if (key != null) {
+            key = trimEnd(key);
+        }
+
+        if (key == null || key.isEmpty() || key.equals(Standard.COMMENT.key()) || key.equals(Standard.HISTORY.key())) {
+            if (value != null) {
+                throw new HeaderCardException("Standard commentary keywords may not have an assigned value.");
             }
+
+            // Force comment
+            type = null;
         }
 
         set(key, value, comment, type);
@@ -543,14 +548,15 @@ public class HeaderCard implements CursorValue<String>, Cloneable {
 
         type = aType;
 
-        // AK: Map null and blank keys to BLANKS.key()
-        // This simplifies things as we won't have to check for null keys separately!
-        if ((aKey == null) || aKey.trim().isEmpty()) {
-            aKey = EMPTY_KEY;
+        // Remove trailing spaces
+        if (aKey != null) {
+            aKey = trimEnd(aKey);
         }
 
-        if (aKey.isEmpty() && aValue != null) {
-            throw new HeaderCardException("Blank or null key for value: [" + sanitize(aValue) + "]");
+        // AK: Map null and blank keys to BLANKS.key()
+        // This simplifies things as we won't have to check for null keys separately!
+        if ((aKey == null) || aKey.isEmpty()) {
+            aKey = EMPTY_KEY;
         }
 
         try {
@@ -772,7 +778,7 @@ public class HeaderCard implements CursorValue<String>, Cloneable {
      * @see    #isCommentStyleCard()
      */
     public synchronized boolean isKeyValuePair() {
-        return !isCommentStyleCard() && !(key.isEmpty() || value == null);
+        return !isCommentStyleCard() && value != null;
     }
 
     /**